+++ tests/testthat/test-SignaturePresence192.R
@@ -1,4 +1,4 @@
-context("SignaturePresenceTest for SBS192")
+context("test-SignaturePresence192.R SignaturePresenceTest1 for SBS192")
 
 test_that("SignaturePresence1 SBS192 1", {
   input <- c(SBS1 = 1000, SBS22 = 2000)
@@ -12,11 +12,13 @@ test_that("SignaturePresence1 SBS192 2", {
   input <- c(SBS5 = 10, SBS1 = 1000, SBS22 = 2000)
   retval <- TestSignaturePresenceTest1(
     sig.counts = input, 
-    input.sigs = PCAWG7::signature$genome$SBS192)
+    input.sigs = PCAWG7::signature$genome$SBS192,
+    eval_f = ObjFnBinomMaxLHMustRound)
   testthat::expect_equal(retval$chisq.p, 0.7378917, tolerance = 1e-5)
 })
 
 test_that("SignaturePresence1 SBS192 3", {
+  testthat::skip_if_not(Sys.getenv("MSIGACT_TEST_LENGTH") == "long")
   input <- c(SBS5 = 500, SBS3 = 10000, SBS22 = 20000)
   retval <- TestSignaturePresenceTest1(
     sig.counts = input, 
@@ -25,6 +27,7 @@ test_that("SignaturePresence1 SBS192 3", {
 })
 
 test_that("SignaturePresence1 SBS192 4", {
+  testthat::skip_if_not(Sys.getenv("MSIGACT_TEST_LENGTH") == "long")
   input <- c(SBS5 = 2000, SBS3 = 10000, SBS22 = 20000)
   retval <- TestSignaturePresenceTest1(
     sig.counts = input, 
@@ -42,6 +45,7 @@ test_that("SignaturePresence1 SBS192 5", {
 })
 
 test_that("SignaturePresence1 SBS192 6", {
+  testthat::skip_if_not(Sys.getenv("MSIGACT_TEST_LENGTH") == "long")
   input <- c(SBS5 = 800, SBS3 = 10000, SBS22 = 20000, SBS40 = 1000)
   retval <- TestSignaturePresenceTest1(
     sig.counts = input,